Property file implementation is completed.
More Details:
As per the new requirement, I started working on implementing property
files for Tapas tool. Accordingly, app will pull get irods properties
from a editable file (tapas.propperties) available to user. This file
will be available to user in usual local cache folder. At any point this
path can be changed as per users will. These properties will be picked
before loading of app and will be constant throughout the user session.
If property file is edited, then new values will be picked in next
immediate session. On the other hand, In case prop file is not
available, then app will pull a backup copy of it with default values
from within itself (jar). As this backup copy is processed, a local copy
of it is left in cache folder, so that user can edit it in future
sessions.